+++ wxcadm/__init__.py
@@ -7,9 +7,12 @@ log.setLevel(logging.DEBUG)
 log.addHandler(logging.NullHandler())
 
 from .webex import Webex
+from .org import Org
 from .xsi import XSIEvents, Call, XSI, XSICallQueue
 from .redsky import RedSky
 from .meraki import Meraki
 from .exceptions import *
 from .common import *
 from .wholesale import Wholesale
+from .location import Location, LocationSchedule
+from .location_features import *